updating build scripts

+++ build-linux.fsx
@@ -93,16 +93,18 @@ let args = fsi.CommandLineArgs |> Array.skip 1  // skip the script name
 // Steps
 // -------------------------------
 
-// 1. Clean the project dir
-log "cleaning directories"
+// 1. Clean the directories
+if args |> Array.contains "--clean" then
+    log "cleaning directories"
+    runProcess "dotnet" "clean" projectDir
 
-let binPath = Path.Combine(projectDir, "bin")
-if Directory.Exists(binPath) then
-    Directory.Delete(binPath, true) 
+    let binPath = Path.Combine(projectDir, "bin")
+    if Directory.Exists(binPath) then
+        Directory.Delete(binPath, true) 
 
-let objPath = Path.Combine(projectDir, "obj")
-if Directory.Exists(objPath) then
-    Directory.Delete(objPath, true) 
+    let objPath = Path.Combine(projectDir, "obj")
+    if Directory.Exists(objPath) then
+        Directory.Delete(objPath, true) 
 
 // 1. Publish the project
 let selfContained =
